Download artifacts with the newest API version
With the previous version we were incurring in pagination limits, this would mean that we had to manage that manually after having 30 jobs running. Which may be very common for WIP PRs.

I implemented the same logic by downloading the artifacts from the job itself in the Rest API v2, which, for the quantity of jobs and artifacts, won't be paginated as there are always a few of them in separate requests.

Also, the pipelines are normally returned by Circle Ci in chronological order from the newest to the oldest, meaning, that even if it is paginated we'll always be geting the most recent.

Finally, this implementation waits for the artifacts job itself to finish and not for the entire workflow as opposed to the previous one.